requirements.txt
Packages:

Use "pip install" to install the packages below:

djangorestframework
django-bootstrap-v5
crispy-bootstrap5
django-crispy-forms
mysqlclient